fix(solana): close host ABI copies and type gate
Regenerate IDL/Codama after the 32-byte KMS id, reject euint160/256 at
the type gate, and store HCU costs as const tables.
