+++ coprocessor/fhevm-engine/host-listener/idl/solana_abi_golden.json
@@ -216,19 +216,19 @@
     },
     {
       "category": "account",
-      "fixture_hex": "439eb0f8ab93a1dc0101010101010101010101010101010101010101010101010101010101010101420000000000000043000000000000000405060708090a0b0c0d0e0f101112131415161705060708090a0b0c0d0e0f101112131415161718060708090a0b0c0d0e0f101112131415161718190708090a0b0c0d0e0f101112131415161718191a08090a0b0c0d0e0f101112131415161718191a1b090a0b0c0d0e0f101112131415161718191a1b1c0a0b0c0d0e0f101112131415161718191a1b1c1d0b0c0d0e0f101112131415161718191a1b1c1d1e0c0d0e0f101112131415161718191a1b1c1d1e1f060708090a0b0c0d0e0f101112131415161718191a1b490000000000000001014c000000000000004d000000000000004e000000000000004f0000000000000010",
-      "fixture_len": 301,
+      "fixture_hex": "439eb0f8ab93a1dc0101010101010101010101010101010101010101010101010101010101010101420000000000000043000000000000000405060708090a0b0c0d0e0f101112131415161705060708090a0b0c0d0e0f101112131415161718060708090a0b0c0d0e0f101112131415161718190708090a0b0c0d0e0f101112131415161718191a08090a0b0c0d0e0f101112131415161718191a1b090a0b0c0d0e0f101112131415161718191a1b1c0a0b0c0d0e0f101112131415161718191a1b1c1d0b0c0d0e0f101112131415161718191a1b1c1d1e0c0d0e0f101112131415161718191a1b1c1d1e1f060708090a0b0c0d0e0f101112131415161718191a1b090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20212223242526272801014c000000000000004d000000000000004e000000000000004f0000000000000010",
+      "fixture_len": 325,
       "name": "HostConfig",
       "program": "zama_host",
-      "schema_hash": "6f51e02f8d16a76d0df5b4b9db885239de9e3c76c4cf40215e6e454f29591d2f"
+      "schema_hash": "0e523818e410c4c459bdf6bc09975ee23546c16ff5460c18bad8fee155afc8fa"
     },
     {
       "category": "account",
-      "fixture_hex": "5d00cd284a3fc99541000000000000000100000002030405060708090a0b0c0d0e0f101112131415030405060105",
-      "fixture_len": 46,
+      "fixture_hex": "5d00cd284a3fc9950102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f200100000002030405060708090a0b0c0d0e0f101112131415030405060105",
+      "fixture_len": 70,
       "name": "KmsContext",
       "program": "zama_host",
-      "schema_hash": "bebe0b86e62aab6999d55e98817c684cab07ac4ac41943ef33f72172190e1323"
+      "schema_hash": "6ac55ac9399ccf706259deb7528c6c8159f60277eee30ddd8ec998c8e7702f38"
     },
     {
       "category": "event",
@@ -256,14 +256,6 @@
     },
     {
       "category": "event",
-      "fixture_hex": "ef91bde5640dcdf101020202020202020202020202020202020202020202020202020202020202020203030303030303030303030303030303030303030303030303030303030303034400000000000000",
-      "fixture_len": 81,
-      "name": "HostConfigInitializedEvent",
-      "program": "zama_host",
-      "schema_hash": "a84db130db85b76ae81653ce9bc31bf6238eabbc052b57e7227a8e90c75b3810"
-    },
-    {
-      "category": "event",
       "fixture_hex": "5bf67764f8ea388a010202020202020202020202020202020202020202020202020202020202020202030303030303030303030303030303030303030303030303030303030303030301014600000000000000470000000000000048000000000000004900000000000000",
       "fixture_len": 107,
       "name": "HostConfigUpdatedEvent",
@@ -272,19 +264,19 @@
     },
     {
       "category": "event",
-      "fixture_hex": "debe8686f59f26f0014200000000000000",
-      "fixture_len": 17,
+      "fixture_hex": "debe8686f59f26f00102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f2021",
+      "fixture_len": 41,
       "name": "KmsContextDestroyedEvent",
       "program": "zama_host",
-      "schema_hash": "053677e75dd128eae05bf9bcb687c1f1a3778cab46bdfd328b395bec1226808f"
+      "schema_hash": "c09d9d9817bd3d8a3464c92367f43e472dae08a56cee813777eb27c105fe3b61"
     },
     {
       "category": "event",
-      "fixture_hex": "2b505658eae8f03f01420000000000000001000000030405060708090a0b0c0d0e0f101112131415160405",
-      "fixture_len": 43,
+      "fixture_hex": "2b505658eae8f03f0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f202101000000030405060708090a0b0c0d0e0f101112131415160405",
+      "fixture_len": 67,
       "name": "NewKmsContextEvent",
       "program": "zama_host",
-      "schema_hash": "3fbd2cb6813ae4a9f64a0abda87f3023591218aa932e299c8358cef3f361a256"
+      "schema_hash": "db22c4c025e0d973a3bc8508872533097f04528c51427021210b30321e218430"
     },
     {
       "category": "event",
@@ -304,11 +296,11 @@
     },
     {
       "category": "instruction_args",
-      "fixture_hex": "66e95717aebae95741000000000000000100000002030405060708090a0b0c0d0e0f10111213141503040506",
-      "fixture_len": 44,
+      "fixture_hex": "66e95717aebae9570102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f200100000002030405060708090a0b0c0d0e0f10111213141503040506",
+      "fixture_len": 68,
       "name": "define_kms_context",
       "program": "zama_host",
-      "schema_hash": "cc8a0482562328c29e2061f8ab265e881996322a7fa0a36b9e806d9380f8072e"
+      "schema_hash": "a81cb40815407ec0aa940ef67435214983eb70d534aec36ab26e835d75c6ecda"
     },
     {
       "category": "instruction_args",
@@ -320,11 +312,11 @@
     },
     {
       "category": "instruction_args",
-      "fixture_hex": "2cb818ea5939e8fb4100000000000000",
-      "fixture_len": 16,
+      "fixture_hex": "2cb818ea5939e8fb0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f20",
+      "fixture_len": 40,
       "name": "destroy_kms_context",
       "program": "zama_host",
-      "schema_hash": "7943b3a5d342398e02b4411ffb919afbd24c5cf81be6c74cfedfd5b8d505034b"
+      "schema_hash": "4f8e91513f08fa8a4d286010e0ac116b3b83d9ac24ce50fc9a3e8e60d50071de"
     },
     {
       "category": "instruction_args",
@@ -340,7 +332,7 @@
       "fixture_len": 90,
       "name": "initialize_host_config",
       "program": "zama_host",
-      "schema_hash": "407b5732f1bfda8b6dfd63a68788602ad0c21f05f47aa3f4166c4ad776bb6372"
+      "schema_hash": "45275bbe5b6962050e54ebfdb9d8cb9d6d927e6f2483749e1896582d99e0841c"
     },
     {
       "category": "instruction_args",
@@ -376,6 +368,14 @@
     },
     {
       "category": "instruction_args",
+      "fixture_hex": "fba300345bc2bb5c0101010101010101010101010101010101010101010101010101010101010101",
+      "fixture_len": 40,
+      "name": "set_admin",
+      "program": "zama_host",
+      "schema_hash": "b898a237e42b5ff83610a13ef68aac13627160d34bf7f18fddce4900cc9316b2"
+    },
+    {
+      "category": "instruction_args",
       "fixture_hex": "261e3b2ce901c2b1010000000102030405060708090a0b0c0d0e0f101112131402",
       "fixture_len": 33,
       "name": "set_coprocessor_signers",
@@ -389,6 +389,14 @@
       "name": "set_deny_subject",
       "program": "zama_host",
       "schema_hash": "ddc2a74840ffcc5eeb2dcb45a7d0981246672811abdb93a9e9225d907e7f7696"
+    },
+    {
+      "category": "instruction_args",
+      "fixture_hex": "9f5dad14119885ea410000000000000002030405060708090a0b0c0d0e0f101112131415030405060708090a0b0c0d0e0f10111213141516",
+      "fixture_len": 56,
+      "name": "set_eip712_domain",
+      "program": "zama_host",
+      "schema_hash": "27093649cf7e2d4b5dec20a7d364df8f24e9aff7167cf4aca0328ca25daf06b7"
     },
     {
       "category": "instruction_args",
