+++ solana/programs/zama-host/README.md
@@ -111,9 +111,8 @@ signer), never a user key and never the bare program id (program ids cannot sign
 enforces `contract_address == compute_subject` (any signer). `user_address` is **not** EVM
 `msg.sender`; binding the attestation to that PDA and checking the attested `user_address` are
 **app policy** (confidential-token checks the attested user
-equals the token account owner), mirroring EVM where `userAddress` is attested but the contract
-decides its meaning. Per-state-account (per-mint) scoping is deliberate and finer-grained than EVM's
-per-contract binding.
+equals the token account owner). Per-state-account (per-mint) scoping is deliberate and finer-grained
+than EVM's per-contract binding.
 
 This mirrors the EVM `InputVerification` coprocessor-threshold model; the gateway counterpart is the
 RFC-021 bytes32 path `InputVerification.verifyProofRequestSolana`. The host-listener reconstruct path
