+++ solana/scripts/check_solana_abi.py
@@ -138,7 +138,9 @@ PINNED_SCHEMAS = [
     ("zama_host", "instruction_args", "destroy_kms_context", True),
     ("zama_host", "instruction_args", "revoke_delegation_for_user_decryption", True),
     ("zama_host", "instruction_args", "revoke_permits", True),
+    ("zama_host", "instruction_args", "set_admin", True),
     ("zama_host", "instruction_args", "set_coprocessor_signers", True),
+    ("zama_host", "instruction_args", "set_eip712_domain", True),
     ("zama_host", "instruction_args", "set_deny_subject", True),
     ("zama_host", "instruction_args", "set_grant_deny_list_enabled", True),
     ("zama_host", "instruction_args", "set_hcu_app_trusted", True),
